+++ Assets/Features/UI Design/Scripts/Runtime/Legacy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3923d2d74c62c8f46b8e802b81c1a324
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
