+++ Assets/Features/UI Design/Scripts/Runtime/Utils/TimeReference.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e121a8569d4946344b20ac74c30d61aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
