+++ Assets/Features/UI Design/Scripts/Runtime/Utils/TimeReferenceHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2db7e271fb21a04bb92ed3f29f8056c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
