+++ Assets/Scripts/TimeDiff919.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 37b2d9942018d10448178eabdbdf6bfd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
